requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
httpx==0.27.0
jinja2==3.1.4
python-dotenv==1.0.1
sentence-transformers==2.6.1
faiss-cpu==1.8.0
rank-bm25==0.2.2
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
pydantic-settings==2.4.0
google-generativeai==0.5.3
python-docx==1.1.0
markdown-it-py==3.0.0
python-multipart==0.0.9
slowapi==0.1.9
pytest==8.2.1
pytest-asyncio==0.23.6
respx==0.21.1
